# Report errors when `usize` overflows.