This makes it now safe to use on 16-bit systems, with one small exception in V2DeflateSerializer.

This necessitated adding a few error types to things that used to overflow. Various other little cleanups performed while following the logic around.

Fixes #32.

The various error types cause a small performance hit for serialization, generally on the order of several percent. Serialization is still really fast, so I'm not too worried about it, but it does sting a bit to watch some hard-won performance slip away. C'est la vie. This isn't the first slow thing we have to do for safety. :'(

master:

```
test serialize_large_dense_v2            ... bench:   7,048,196 ns/iter (+/- 219,608)
test serialize_large_dense_v2_deflate    ... bench:  96,084,888 ns/iter (+/- 2,199,941)
test serialize_large_sparse_v2           ... bench:   6,397,566 ns/iter (+/- 333,578)
test serialize_large_sparse_v2_deflate   ... bench:  62,400,106 ns/iter (+/- 891,614)
test serialize_medium_dense_v2           ... bench:      43,453 ns/iter (+/- 1,176)
test serialize_medium_sparse_v2          ... bench:      39,222 ns/iter (+/- 658)
test serialize_small_dense_v2            ... bench:       5,100 ns/iter (+/- 125)
test serialize_small_sparse_v2           ... bench:       1,889 ns/iter (+/- 50)
test serialize_tiny_dense_v2             ... bench:       1,239 ns/iter (+/- 68)
test serialize_tiny_sparse_v2            ... bench:         596 ns/iter (+/- 39)
```

This PR:

```
test serialize_large_dense_v2            ... bench:   7,150,907 ns/iter (+/- 206,409)
test serialize_large_dense_v2_deflate    ... bench:  94,711,640 ns/iter (+/- 4,685,772)
test serialize_large_sparse_v2           ... bench:   6,591,700 ns/iter (+/- 181,256)
test serialize_large_sparse_v2_deflate   ... bench:  63,085,754 ns/iter (+/- 842,193)
test serialize_medium_dense_v2           ... bench:      47,104 ns/iter (+/- 1,768)
test serialize_medium_sparse_v2          ... bench:      41,334 ns/iter (+/- 974)
test serialize_small_dense_v2            ... bench:       6,084 ns/iter (+/- 121)
test serialize_small_sparse_v2           ... bench:       1,923 ns/iter (+/- 47)
test serialize_tiny_dense_v2             ... bench:       1,338 ns/iter (+/- 43)
test serialize_tiny_sparse_v2            ... bench:         588 ns/iter (+/- 13)
```