@jonhoo 2017-04-21
Sorry, extremely busy for an upcoming deadline. Looks fine, though I'm a bit sad that we now need to carry more `Result`/`Option` wrappers around.