@marshallpierce 2017-04-21
No problem. (FYI I'll be moving soon so I'll probably drop off the radar for a while too.) Java gets to avoid this logic because arrays are always guaranteed to have 32 bit indexes. :/